+++ src/main/java/com/appcenter/practice/service/TodoService.java
@@ -2,12 +2,14 @@ package com.appcenter.practice.service;
 
 
 import com.appcenter.practice.domain.Bucket;
+import com.appcenter.practice.domain.Member;
 import com.appcenter.practice.domain.Todo;
 import com.appcenter.practice.dto.request.todo.AddTodoReq;
 import com.appcenter.practice.dto.request.todo.UpdateTodoReq;
 import com.appcenter.practice.dto.response.todo.TodoRes;
 import com.appcenter.practice.exception.CustomException;
 import com.appcenter.practice.repository.BucketRepository;
+import com.appcenter.practice.repository.MemberRepository;
 import com.appcenter.practice.repository.TodoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,8 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.appcenter.practice.common.StatusCode.BUCKET_NOT_EXIST;
-import static com.appcenter.practice.common.StatusCode.TODO_NOT_EXIST;
+import static com.appcenter.practice.common.StatusCode.*;
 
 @Service
 @Transactional(readOnly = true)
@@ -25,6 +26,7 @@ import static com.appcenter.practice.common.StatusCode.TODO_NOT_EXIST;
 public class TodoService {
     private final TodoRepository todoRepository;
     private final BucketRepository bucketRepository;
+    private final MemberRepository memberRepository;
 
 
     public List<TodoRes> getTodoList(Long bucketId){
@@ -37,32 +39,46 @@ public class TodoService {
 
 
     @Transactional
-    public TodoRes saveTodo(Long bucketId, AddTodoReq reqDto){
+    public TodoRes saveTodo(Long memberId,Long bucketId, AddTodoReq reqDto){
+        Member member=findByMemberId(memberId);
         Bucket bucket=findByBucketId(bucketId);
-        Todo todo =todoRepository.save(reqDto.toEntity(bucket));
-        return TodoRes.from(todo);
+        if(member.equals(bucket.getMember())){
+            Todo todo =todoRepository.save(reqDto.toEntity(bucket));
+            return TodoRes.from(todo);
+        }
+        else throw new CustomException(AUTHORIZATION_INVALID);
     }
 
     @Transactional
-    public TodoRes updateTodo(Long todoId, UpdateTodoReq reqDto){
+    public TodoRes updateTodo(Long memberId, Long todoId, UpdateTodoReq reqDto){
+        Member member=findByMemberId(memberId);
         Todo todo =findByTodoId(todoId);
-        todo.changeContent(reqDto.getContent());
+        if(member.equals(todo.getBucket().getMember())){
+            todo.changeContent(reqDto.getContent());
+            return TodoRes.from(todo);
+        }
+        else throw new CustomException(AUTHORIZATION_INVALID);
 
-        return TodoRes.from(todo);
     }
 
     @Transactional
-    public TodoRes completeTodo(Long todoId){
+    public TodoRes completeTodo(Long memberId, Long todoId){
+        Member member=findByMemberId(memberId);
         Todo todo=findByTodoId(todoId);
-        todo.changeCompleted();
-
-        return TodoRes.from(todo);
+        if(member.equals(todo.getBucket().getMember())){
+            todo.changeCompleted();
+            return TodoRes.from(todo);
+        }
+        else throw new CustomException(AUTHORIZATION_INVALID);
     }
 
     @Transactional
-    public void deleteTodo(Long todoId){
+    public void deleteTodo(Long memberId, Long todoId){
+        Member member=findByMemberId(memberId);
         Todo todo =findByTodoId(todoId);
-        todoRepository.deleteById(todo.getId());
+        if(member.equals(todo.getBucket().getMember()))
+            todoRepository.deleteById(todo.getId());
+        else throw new CustomException(AUTHORIZATION_INVALID);
     }
 
     private Todo findByTodoId(Long todoId) {
@@ -73,6 +89,11 @@ public class TodoService {
     private Bucket findByBucketId(Long bucketId){
         return bucketRepository.findById(bucketId)
                 .orElseThrow(()->new CustomException(BUCKET_NOT_EXIST));
+    }
+
+    private Member findByMemberId(Long memberId){
+        return memberRepository.findById(memberId)
+                .orElseThrow(()->new CustomException(MEMBER_NOT_EXIST));
     }
 
 }
